requirements.txt
numpy == 1.26.3 
pandas == 2.1.0
streamlit == 1.43.2
python-jobspy == 1.1.78